Rename graphql_utils.js to graphql_operations.js
