+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Changed
 - Moved gmail to common folder
 - In, commonGraphlValidation() added expect(response.body).to.not.equal('OK')
+- Rename graphql_utils.js to graphql_operations.js
 
 ### Added
 - Removed delayBetweenRetries + Added reload then fillContact info first then shipping info
