+++ cypress-shared/support/b2b/common.js
@@ -1,7 +1,7 @@
 import selectors from '../common/selectors.js'
 import { generateEmailWithSuffix, validateToastMsg } from './utils.js'
 import { BUTTON_LABEL, TOAST_MSG } from '../validation_text.js'
-import { GRAPHL_OPERATIONS } from '../graphql_utils.js'
+import { GRAPHL_OPERATIONS } from '../graphql_operations.js'
 import { updateRetry } from '../common/support.js'
 
 // Define constants
