+++ cypress-shared/support/b2b/organization_request.js
@@ -1,6 +1,6 @@
 import selectors from '../common/selectors.js'
 import { OTHER_ROLES } from './utils.js'
-import { GRAPHL_OPERATIONS } from '../graphql_utils.js'
+import { GRAPHL_OPERATIONS } from '../graphql_operations.js'
 import { BUTTON_LABEL } from '../validation_text.js'
 import { FAIL_ON_STATUS_CODE } from '../common/constants.js'
 import { OrganizationRequestStatus } from './constants.js'
