+++ cypress-shared/support/graphql_operations.js
